Final version of baseline model
- Also updated README.md
- Cleaned the repo
