+++ requirements.txt
@@ -1,4 +1,3 @@
 torch=2.1.0
 numpy=1.26.1
 matplotlib=3.8.0
-torchvision=0.16.0
